Update README.md
A wafer is the foundational material used in the manufacturing of microelectronic devices. It is a thin, circular slice of a crystalline semiconductor material such as silicon, gallium arsenide, or silicon carbide. Wafers are integral to the semiconductor industry and serve as the platform for creating integrated circuits (ICs), which are the brains of modern electronic devices.